% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vindex.R
\name{CM}
\alias{CM}
\title{Cálculo del Índice de Centro de Masa}
\usage{
CM(Sv_mean, Depth_mean)
}
\arguments{
\item{Sv_mean}{vector con los valores de Sv promedio}

\item{Depth_mean}{vector con los valores de prodfundidad promedio}
}
\description{
Cálculo de índice de centro de masa para los cardúmenes.
}
