% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_reg.R
\name{filter.reg}
\alias{filter.reg}
\title{Limpieza de datos en la bitácora de regiones}
\usage{
filter.reg(file, sp = c("ANC", "MUN", "JUR",
            "VIN", "POT", "BAG", "CAB", "SAM"), frequency,
            path.save)
}
\arguments{
\item{file}{bitácora de regiones en formato \emph{.csv}}

\item{sp}{especies a analizar. Por default se consideran: anchoveta
(ANC), munida (MUN), jurel (JUR), vinciguerria (VIN), pota (POT),
bagre (BAG), caballa (CAB) y samasa (SAM).}

\item{frequency}{frecuencia a la que corresponden los datos.}

\item{path.save}{ruta donde será guardada la nueva base de datos.}
}
\description{
Función para filtrar la base de datos de regiones teniendo en
cuenta los errores producidos durante la medición de los datos.
}
\details{
Los filtros se realizan en dos niveles: generales y por especie.

En los filtros generales se corrigen todos los valores que por default
produce echoview para las medidas de longitud, latitud y Sv.

Para evitar malas detecciones dentro del campo cercano del transductor, se
considera el campo cercano de cada frecuencia y se eliminan las regiones
que se encuentran dentro de la mismas.

En los filtros por especie consideran el valor de Sv por especie
y la profundidad máxima a la cual pueden llegar.
}
