% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vindex.R
\name{IA}
\alias{IA}
\title{Cálculo del Índice de Agregación}
\usage{
IA(Sv_mean, Depth_mean)
}
\arguments{
\item{Sv_mean}{vector con los valores de Sv promedio}

\item{Depth_mean}{vector con los valores de prodfundidad promedio}
}
\description{
Cálculo de índice de agregación para los cardúmenes.
}
