% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vindex.R
\name{Inertia}
\alias{Inertia}
\title{Cálculo de Inercia}
\usage{
Inertia(Depth_mean, CM, Sv_mean)
}
\arguments{
\item{Depth_mean}{vector con los valores de prodfundidad promedio}

\item{CM}{vector con el valor del centro de masa}

\item{Sv_mean}{vector con los valores de Sv promedio}
}
\description{
Cálculo de inercia vertical para los cardúmenes.
}
